Changed random_num back to ints and changed weather data types to ints

+++ week3/d01/Assignment_3_1_2/includes/utils.h
@@ -1,7 +1,7 @@
 #ifndef UTILS_H
 #define UTILS_H
 
-double random_num(double a, double b);
+int random_num(int a, int b);
 void clear_cin();
 
 #endif
